vif_plug_vhostuser_fp/common: support containerized setups
In case of containerized setup like RedHat OSP, nova-compute, libvirtd,
virtual-accelerator can be installed and run in different containers.
Hence, we can't simply call fp-vdev to create/delete vhostuser port,
because fp-vdev can be absent in nova-compute container. As workaround,
virtual-accelerator package brings and starts fp-rpcd daemon. It serves
as a proxy to bypass fp-vdev requests from containers to fp-vdev binary
in a host-hypervisor. fp-rpcd.sock should be created in a folder shared
between nova_compute and the host-hypervisor.

os-vif-plugin-vhostuser-fp plugin package is now depending from
fp-vdev-remote module, which brings fp-vdev-remote client to fp-rpcd.
fp-vdev-remote in nova_compute container connects to fp-rpcd.sock in the
shared folder and sends virtual port create/delete requests to fp-vdev,
which is in the host-hypervisor.

Moreover, vhostuser socket should be also created in a folder shared
between nova_libvirt container and host-hypervisor, because qemu process
is launching in nova_libvirtd container while creating a VM, but fp-rte
is always running in the hosti-hypervisor.

To keep compatibility with any fastpath version and limit modifications,
a command to create or delete a virtual port is identical whatever the
binary is used (fp-vdev or fp-vdev-remote). Choice of a proper 'fp-vdev'
binary is done in fp-vdev-remote module by checking of presence of the
fp-rpcd.sock file.

Signed-off-by: Valentina Krasnobaeva <valentina.krasnobaeva@6wind.com>
Acked-by: Maxime Leroy <maxime.leroy@6wind.com>